INW 128 main fragment code removed
