+++ Android/app/src/main/java/ru/inwords/inwords/dagger/AppComponent.kt
@@ -7,7 +7,6 @@ import ru.inwords.inwords.authorisation.di.AuthorisationDaggerModule
 import ru.inwords.inwords.data.source.database.RoomTypeConverter
 import ru.inwords.inwords.game.data.deferred.level_score.LevelScoreUploadWorker
 import ru.inwords.inwords.game.di.OctoGameDaggerModule
-import ru.inwords.inwords.home.HomeFragmentDaggerModule
 import ru.inwords.inwords.main_activity.di.MainActivityDaggerModule
 import ru.inwords.inwords.policy.di.PolicyFragmentDaggerModule
 import ru.inwords.inwords.profile.di.ProfileFragmentDaggerModule
@@ -17,7 +16,6 @@ import javax.inject.Singleton
 @Singleton
 @Component(modules = [AppModule::class,
     AuthorisationDaggerModule::class,
-    HomeFragmentDaggerModule::class,
     TranslationDaggerModule::class,
     OctoGameDaggerModule::class,
     MainActivityDaggerModule::class,
