--- Android/app/src/main/java/ru/inwords/inwords/home/HomeFragment.kt
@@ -1,87 +1,0 @@
-package ru.inwords.inwords.home
-
-import android.os.Bundle
-import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.google.android.material.snackbar.Snackbar
-import io.reactivex.disposables.Disposable
-import ru.inwords.inwords.R
-import ru.inwords.inwords.core.recycler.EmptySpaceItemDecoration
-import ru.inwords.inwords.core.recycler.fixOverscrollBehaviour
-import ru.inwords.inwords.core.rxjava.SchedulersFacade
-import ru.inwords.inwords.core.utils.observe
-import ru.inwords.inwords.databinding.FragmentHomeBinding
-import ru.inwords.inwords.home.recycler.CardsRecyclerAdapter
-import ru.inwords.inwords.presentation.view_scenario.FragmentWithViewModelAndNav
-
-
-class HomeFragment : FragmentWithViewModelAndNav<HomeViewModel, HomeViewModelFactory, FragmentHomeBinding>() {
-    override val layout = R.layout.fragment_home
-    override val classType = HomeViewModel::class.java
-
-    override fun bindingInflate(inflater: LayoutInflater, container: ViewGroup?, attachToRoot: Boolean): FragmentHomeBinding {
-        return FragmentHomeBinding.inflate(inflater, container, attachToRoot)
-    }
-
-    private lateinit var recyclerAdapter: CardsRecyclerAdapter
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        setupToolbar()
-
-        observeData(view)
-
-        setupRecycler()
-        subscribeRecycler().disposeOnViewDestroyed()
-    }
-
-    private fun observeData(view: View) {
-        observe(viewModel.error) {
-            Snackbar.make(view, R.string.unable_to_load_exercise, Snackbar.LENGTH_SHORT).show()
-        }
-
-        observe(viewModel.profile) {
-            binding.toolbar.title = it.nickName
-        }
-    }
-
-    private fun setupToolbar() {
-        binding.toolbar.setOnMenuItemClickListener { menuItem ->
-            when (menuItem.itemId) {
-                R.id.action_profile -> {
-                    viewModel.navigateToProfile()
-                    true
-                }
-                else -> false
-            }
-        }
-    }
-
-    private fun setupRecycler() {
-        recyclerAdapter = CardsRecyclerAdapter { viewModel.handleNavigation(it) }
-
-        val dividerItemDecoration = EmptySpaceItemDecoration(resources.getDimensionPixelSize(R.dimen.space_medium))
-
-        with(binding.cardsRecycler) {
-            layoutManager = LinearLayoutManager(context)
-            addItemDecoration(dividerItemDecoration)
-            adapter = recyclerAdapter
-        }
-    }
-
-    private fun subscribeRecycler(): Disposable {
-        return viewModel.cardWrappers
-            .observeOn(SchedulersFacade.ui())
-            .subscribe({ result ->
-                recyclerAdapter.accept(result)
-
-                fixOverscrollBehaviour(binding.cardsRecycler)
-            }) {
-                Log.e(javaClass.simpleName, it.message.orEmpty())
-            }
-    }
-}
