--- Android/app/src/main/java/ru/inwords/inwords/home/HomeFragmentDaggerModule.kt
@@ -1,12 +1,0 @@
-package ru.inwords.inwords.home
-
-
-import dagger.Module
-import dagger.android.ContributesAndroidInjector
-
-@Module
-abstract class HomeFragmentDaggerModule {
-    @ContributesAndroidInjector
-    internal abstract fun mainFragmentInjector(): HomeFragment
-}
-
