--- Android/app/src/main/java/ru/inwords/inwords/home/HomeViewModelFactory.kt
@@ -1,23 +1,0 @@
-package ru.inwords.inwords.home
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import ru.inwords.inwords.profile.domain.interactor.ProfileInteractor
-import ru.inwords.inwords.translation.domain.interactor.TranslationWordsInteractor
-import javax.inject.Inject
-
-class HomeViewModelFactory @Inject
-internal constructor(
-    private val translationWordsInteractor: TranslationWordsInteractor,
-    private val profileInteractor: ProfileInteractor
-) : ViewModelProvider.Factory {
-
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(HomeViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return HomeViewModel(translationWordsInteractor, profileInteractor) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-
-}
