--- Android/app/src/main/res/layout/card_dictionary.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/cardView"
-    android:layout_width="match_parent"
-    android:layout_height="@dimen/default_card_height"
-    android:clickable="true"
-    android:focusable="true"
-    android:foreground="?android:attr/selectableItemBackground"
-    app:cardCornerRadius="@dimen/card_corner_radius"
-    app:cardElevation="@dimen/elevation_medium">
-
-    <TextView
-        android:id="@+id/title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/space_medium"
-        android:layout_marginTop="@dimen/space_xsmall"
-        android:layout_marginEnd="@dimen/space_giantxx"
-        android:elevation="@dimen/elevation_medium"
-        android:ellipsize="end"
-        android:lines="1"
-        android:text="@string/dictionary_title"
-        android:textAppearance="@style/TextAppearance.Inwords.Headline5.Default" />
-
-    <TextView
-        android:id="@+id/dictSize"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        android:layout_marginStart="@dimen/space_medium"
-        android:layout_marginEnd="@dimen/space_giantxx"
-        android:layout_marginBottom="@dimen/space_xsmall"
-        android:elevation="@dimen/elevation_medium"
-        android:ellipsize="end"
-        android:lines="1"
-        android:textAppearance="@style/TextAppearance.Inwords.Body1.Primary"
-        tools:text="Слов в словаре" />
-
-    <ImageView
-        android:id="@+id/dict"
-        android:layout_width="@dimen/default_card_height"
-        android:layout_height="@dimen/default_card_height"
-        android:layout_gravity="end"
-        android:contentDescription="@string/avatar"
-        android:elevation="@dimen/elevation_medium"
-        android:src="@drawable/ic_dictionary_layered" />
-
-</androidx.cardview.widget.CardView>
