--- Android/app/src/main/res/layout/card_profile.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/cardView"
-    android:layout_width="match_parent"
-    android:layout_height="@dimen/default_card_height"
-    android:clickable="true"
-    android:focusable="true"
-    android:foreground="?android:attr/selectableItemBackground"
-    app:cardCornerRadius="@dimen/card_corner_radius"
-    app:cardElevation="@dimen/elevation_medium">
-
-    <TextView
-        android:id="@+id/name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/space_giantx"
-        android:layout_marginTop="@dimen/space_xsmall"
-        android:elevation="@dimen/elevation_medium"
-        android:ellipsize="end"
-        android:lines="1"
-        android:paddingStart="@dimen/space_xsmall"
-        android:paddingEnd="@dimen/space_xsmall"
-        android:textAppearance="@style/TextAppearance.Inwords.Headline5.Default"
-        tools:text="Любитель Языков" />
-
-    <TextView
-        android:id="@+id/experience"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        android:layout_marginStart="@dimen/space_giantx"
-        android:layout_marginBottom="@dimen/space_xsmall"
-        android:elevation="@dimen/elevation_medium"
-        android:ellipsize="end"
-        android:maxLines="1"
-        android:paddingStart="@dimen/space_xsmall"
-        android:paddingEnd="@dimen/space_xsmall"
-        android:textAppearance="@style/TextAppearance.Inwords.Body1.Primary"
-        tools:text="@string/user_experience" />
-
-    <com.facebook.drawee.view.SimpleDraweeView
-        android:id="@+id/avatar"
-        android:layout_width="@dimen/default_card_height"
-        android:layout_height="@dimen/default_card_height"
-        android:contentDescription="@string/avatar"
-        android:elevation="@dimen/elevation_medium"
-        app:actualImageScaleType="fitCenter"
-        app:failureImage="@drawable/octopus_welcome"
-        app:failureImageScaleType="fitCenter" />
-</androidx.cardview.widget.CardView>
