--- Android/app/src/main/res/layout/card_profile_shimmer.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.facebook.shimmer.ShimmerFrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    app:shimmer_auto_start="false">
-
-    <androidx.cardview.widget.CardView
-        android:id="@+id/cardView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:cardCornerRadius="@dimen/card_corner_radius"
-        app:cardElevation="@dimen/elevation_medium">
-
-        <View
-            android:id="@+id/avatar"
-            android:layout_width="@dimen/default_card_height"
-            android:layout_height="@dimen/default_card_height"
-            android:background="@drawable/rounded_rect"
-            android:backgroundTint="@color/shimmerColor"
-            android:contentDescription="@string/avatar"
-            android:elevation="@dimen/elevation_medium" />
-
-        <View
-            android:id="@+id/name"
-            android:layout_width="wrap_content"
-            android:layout_height="@dimen/text_24sp_box"
-            android:layout_marginStart="@dimen/space_giantxx"
-            android:layout_marginTop="@dimen/space_xsmall"
-            android:layout_marginEnd="@dimen/space_xsmall"
-            android:background="@drawable/rounded_rect"
-            android:backgroundTint="@color/shimmerColor"
-            android:elevation="@dimen/elevation_medium" />
-
-        <View
-            android:id="@+id/experience"
-            android:layout_width="wrap_content"
-            android:layout_height="@dimen/text_16sp_box"
-            android:layout_gravity="bottom"
-            android:layout_marginStart="@dimen/space_giantxx"
-            android:layout_marginEnd="@dimen/space_xsmall"
-            android:layout_marginBottom="@dimen/space_xsmall"
-            android:background="@drawable/rounded_rect"
-            android:backgroundTint="@color/shimmerColor"
-            android:elevation="@dimen/elevation_medium" />
-    </androidx.cardview.widget.CardView>
-</com.facebook.shimmer.ShimmerFrameLayout>
