+++ Android/app/src/main/res/layout/fragment_policy.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".home.HomeFragment">
+    tools:context=".policy.presentation.PolicyFragment">
 
     <ScrollView
         android:layout_width="match_parent"
