+++ Android/app/src/main/res/values/strings.xml
@@ -1,11 +1,8 @@
 <resources>
     <string name="app_name">InWords</string>
-    <string name="title_home">Главная</string>
     <string name="title_dictionary">Словарь</string>
     <string name="title_game">Тренировка</string>
     <string name="title_profile">Профиль</string>
-
-    <string name="sign_success">Успешно</string>
 
     <!-- sign in start -->
     <string name="sign_in">Вход</string>
@@ -27,6 +24,7 @@
     <string name="sign_up_enter">Создать аккаунт</string>
     <string name="sign_up_have_account">Уже есть аккаунт?</string>
     <string name="sign_up_sign_in">Войти</string>
+    <string name="sign_up_proposal">Создайте аккаунт, чтобы использовать больше возможностей приложения</string>
     <!-- sign up end -->
 
     <!-- password recovery part1 start -->
@@ -51,8 +49,6 @@
     <string name="word_addition_translation">Перевод</string>
     <string name="button_add_text">Добавить</string>
     <string name="button_edit_text">Изменить</string>
-    <string name="sign_in_in_process">Авторизация</string>
-    <string name="sign_up_in_process">Создание аккаунта</string>
     <string name="inwords_logotype">InWords логотип</string>
     <string name="voice_translation">Озвучить перевод</string>
     <string name="yandex_dictionary">Реализовано с помощью сервиса «Яндекс.Словарь</string>
@@ -61,7 +57,15 @@
     <string name="undo_translation_deletion">ОТМЕНИТЬ</string>
     <string name="translation_deleted">Слово удалено</string>
     <string name="translation_multiple_deleted">Слов удалено: %d</string>
-    <string name="dictionary_screen_title">Ваш словарь</string>
+
+    <plurals name="words_in_dictionary">
+        <item quantity="zero">В словаре %1$d слов</item>
+        <item quantity="one">В словаре %1$d слово</item>
+        <item quantity="two">В словаре %1$d слова</item>
+        <item quantity="few">В словаре %1$d слова</item>
+        <item quantity="many">В словаре %1$d слов</item>
+        <item quantity="other">В словаре %1$d слов</item>
+    </plurals>
     <!-- word addition end -->
 
     <!-- gameResource start -->
@@ -78,8 +82,6 @@
     <string name="avatar">Аватар</string>
     <string name="dictionary_title">Ваш словарь</string>
     <string name="error_text_placeholder">Не удалось загрузить данные</string>
-    <string name="loading_text_placeholder">Загрузка профиля</string>
-    <string name="sign_up_proposal">Создайте аккаунт, чтобы использовать больше возможностей приложения</string>
     <string name="user_experience">Опыт на сегодня: %1$d</string>
     <string name="choose_game">Выберите тему</string>
     <string name="game_remove">Удалить</string>
@@ -91,15 +93,6 @@
     <string name="unable_to_load_voice">Не удалось загрузить озвучку</string>
     <string name="unable_to_load_exercise">Не удалось загрузить упражнение</string>
     <string name="exercise_loading">Упражнение загружается</string>
-
-    <plurals name="words_in_dictionary">
-        <item quantity="zero">В словаре %1$d слов</item>
-        <item quantity="one">В словаре %1$d слово</item>
-        <item quantity="two">В словаре %1$d слова</item>
-        <item quantity="few">В словаре %1$d слова</item>
-        <item quantity="many">В словаре %1$d слов</item>
-        <item quantity="other">В словаре %1$d слов</item>
-    </plurals>
     <!-- home end -->
 
     <!-- profile start -->
